Module wasi_snapshot_preview1::bindings::wasi::filesystem::types

Structs

DescriptorFlags

Descriptor flags.

DescriptorStat

File attributes.

DirectoryEntry

A directory entry.

MetadataHashValue

A 128-bit hash value, split into parts because wasm doesn’t have a 128-bit integer type.

Modes

Permissions mode used by open-at, change-file-permissions-at, and similar.

OpenFlags

Open flags used by open-at.

PathFlags

Flags determining the method of how paths are resolved.

Enums

AccessType

Access type used by access-at.

Advice

File or memory access pattern advisory information.

DescriptorType

The type of a filesystem object referenced by a descriptor.

ErrorCode

Error codes returned by functions, similar to errno in POSIX. Not all of these error codes are returned by the functions provided by this API; some are used in higher-level library layers, and others are provided merely for alignment with POSIX.

NewTimestamp

When setting a timestamp, this gives the value to set it to.

Functions

access_at

Check accessibility of a filesystem path.

advise

Provide file advisory information on a descriptor.

append_via_stream

Return a stream for appending to a file, if available.

change_directory_permissions_at

Change the permissions of a directory.

change_file_permissions_at

Change the permissions of a filesystem object that is not a directory.

create_directory_at

Create a directory.

drop_descriptor

Dispose of the specified descriptor, after which it may no longer be used.

drop_directory_entry_stream

Dispose of the specified directory-entry-stream, after which it may no longer be used.

get_flags

Get flags associated with a descriptor.

get_type

Get the dynamic type of a descriptor.

is_same_object

Test whether two descriptors refer to the same filesystem object.

link_at

Create a hard link.

lock_exclusive

Request an exclusive advisory lock for an open file.

lock_shared

Request a shared advisory lock for an open file.

metadata_hash

Return a hash of the metadata associated with a filesystem object referred to by a descriptor.

metadata_hash_at

Return a hash of the metadata associated with a filesystem object referred to by a directory descriptor and a relative path.

open_at

Open a file or directory.

read

Read from a descriptor, without using and updating the descriptor’s offset.

read_directory

Read directory entries from a directory.

read_directory_entry

Read a single directory entry from a directory-entry-stream.

read_via_stream

Return a stream for reading from a file, if available.

readlink_at

Read the contents of a symbolic link.

remove_directory_at

Remove a directory.

rename_at

Rename a filesystem object.

set_size

Adjust the size of an open file. If this increases the file’s size, the extra bytes are filled with zeros.

set_times

Adjust the timestamps of an open file or directory.

set_times_at

Adjust the timestamps of a file or directory.

stat

Return the attributes of an open file or directory.

stat_at

Return the attributes of a file or directory.

symlink_at

Create a symbolic link (also known as a “symlink”).

sync

Synchronize the data and metadata of a file to disk.

sync_data

Synchronize the data of a file to disk.

try_lock_exclusive

Request an exclusive advisory lock for an open file.

try_lock_shared

Request a shared advisory lock for an open file.

unlink_file_at

Unlink a filesystem object that is not a directory.

unlock

Release a shared or exclusive lock on an open file.

write

Write to a descriptor, without using and updating the descriptor’s offset.

write_via_stream

Return a stream for writing to a file, if available.

Type Definitions

Datetime

Descriptor

A descriptor is a reference to a filesystem object, which may be a file, directory, named pipe, special file, or other object on which filesystem calls may be made.

DirectoryEntryStream

A stream of directory entries.

Filesize

File size or length of a region within a file.

InputStream

LinkCount

Number of hard links to an inode.

OutputStream