Enum wasi_snapshot_preview1::bindings::wasi::filesystem::types::ErrorCode

#[repr(u8)]
pub enum ErrorCode {

    Access = 0,
    WouldBlock = 1,
    Already = 2,
    BadDescriptor = 3,
    Busy = 4,
    Deadlock = 5,
    Quota = 6,
    Exist = 7,
    FileTooLarge = 8,
    IllegalByteSequence = 9,
    InProgress = 10,
    Interrupted = 11,
    Invalid = 12,
    Io = 13,
    IsDirectory = 14,
    Loop = 15,
    TooManyLinks = 16,
    MessageSize = 17,
    NameTooLong = 18,
    NoDevice = 19,
    NoEntry = 20,
    NoLock = 21,
    InsufficientMemory = 22,
    InsufficientSpace = 23,
    NotDirectory = 24,
    NotEmpty = 25,
    NotRecoverable = 26,
    Unsupported = 27,
    NoTty = 28,
    NoSuchDevice = 29,
    Overflow = 30,
    NotPermitted = 31,
    Pipe = 32,
    ReadOnly = 33,
    InvalidSeek = 34,
    TextFileBusy = 35,
    CrossDevice = 36,

}

Error codes returned by functions, similar to errno in POSIX. Not all of these error codes are returned by the functions provided by this API; some are used in higher-level library layers, and others are provided merely for alignment with POSIX.

Variants

Access = 0

Permission denied, similar to EACCES in POSIX.

WouldBlock = 1

Resource unavailable, or operation would block, similar to EAGAIN and EWOULDBLOCK in POSIX.

Already = 2

Connection already in progress, similar to EALREADY in POSIX.

BadDescriptor = 3

Bad descriptor, similar to EBADF in POSIX.

Busy = 4

Device or resource busy, similar to EBUSY in POSIX.

Deadlock = 5

Resource deadlock would occur, similar to EDEADLK in POSIX.

Quota = 6

Storage quota exceeded, similar to EDQUOT in POSIX.

Exist = 7

File exists, similar to EEXIST in POSIX.

FileTooLarge = 8

File too large, similar to EFBIG in POSIX.

IllegalByteSequence = 9

Illegal byte sequence, similar to EILSEQ in POSIX.

InProgress = 10

Operation in progress, similar to EINPROGRESS in POSIX.

Interrupted = 11

Interrupted function, similar to EINTR in POSIX.

Invalid = 12

Invalid argument, similar to EINVAL in POSIX.

Io = 13

I/O error, similar to EIO in POSIX.

IsDirectory = 14

Is a directory, similar to EISDIR in POSIX.

Loop = 15

Too many levels of symbolic links, similar to ELOOP in POSIX.

TooManyLinks = 16

Too many links, similar to EMLINK in POSIX.

MessageSize = 17

Message too large, similar to EMSGSIZE in POSIX.

NameTooLong = 18

Filename too long, similar to ENAMETOOLONG in POSIX.

NoDevice = 19

No such device, similar to ENODEV in POSIX.

NoEntry = 20

No such file or directory, similar to ENOENT in POSIX.

NoLock = 21

No locks available, similar to ENOLCK in POSIX.

InsufficientMemory = 22

Not enough space, similar to ENOMEM in POSIX.

InsufficientSpace = 23

No space left on device, similar to ENOSPC in POSIX.

NotDirectory = 24

Not a directory or a symbolic link to a directory, similar to ENOTDIR in POSIX.

NotEmpty = 25

Directory not empty, similar to ENOTEMPTY in POSIX.

NotRecoverable = 26

State not recoverable, similar to ENOTRECOVERABLE in POSIX.

Unsupported = 27

Not supported, similar to ENOTSUP and ENOSYS in POSIX.

NoTty = 28

Inappropriate I/O control operation, similar to ENOTTY in POSIX.

NoSuchDevice = 29

No such device or address, similar to ENXIO in POSIX.

Overflow = 30

Value too large to be stored in data type, similar to EOVERFLOW in POSIX.

NotPermitted = 31

Operation not permitted, similar to EPERM in POSIX.

Pipe = 32

Broken pipe, similar to EPIPE in POSIX.

ReadOnly = 33

Read-only file system, similar to EROFS in POSIX.

InvalidSeek = 34

Invalid seek, similar to ESPIPE in POSIX.

TextFileBusy = 35

Text file busy, similar to ETXTBSY in POSIX.

CrossDevice = 36

Cross-device link, similar to EXDEV in POSIX.

Implementations

impl ErrorCode

pub fn name(&self) -> &'static str

pub fn message(&self) -> &'static str