Function wasi_snapshot_preview1::bindings::wasi::io::streams::read

pub fn read(
    this: InputStream,
    len: u64
) -> Result<(Vec<u8>, StreamStatus), StreamError>

Perform a non-blocking read from the stream.

This function returns a list of bytes containing the data that was read, along with a stream-status which, indicates whether further reads are expected to produce data. The returned list will contain up to len bytes; it may return fewer than requested, but not more.

Once a stream has reached the end, subsequent calls to read or skip will always report end-of-stream rather than producing more data.

If len is 0, it represents a request to read 0 bytes, which should always succeed, assuming the stream hasn’t reached its end yet, and return an empty list.

The len here is a u64, but some callees may not be able to allocate a buffer as large as that would imply. FIXME: describe what happens if allocation fails.

When the returned stream-status is open, the length of the returned value may be less than len. When an empty list is returned, this indicates that no more bytes were available from the stream at that time. In that case the subscribe-to-input-stream pollable will indicate when additional bytes are available for reading.