Module wasi_snapshot_preview1::bindings::wasi::sockets::ip_name_lookup

Functions

drop_resolve_address_stream

Dispose of the specified resolve-address-stream, after which it may no longer be used.

resolve_addresses

Resolve an internet host name to a list of IP addresses.

resolve_next_address

Returns the next address from the resolver.

subscribe

Create a pollable which will resolve once the stream is ready for I/O.

Type Definitions

ErrorCode

IpAddress

IpAddressFamily

Network

Pollable

ResolveAddressStream