Function wasi_snapshot_preview1::bindings::wasi::sockets::tcp::shutdown

pub fn shutdown(
    this: TcpSocket,
    shutdown_type: ShutdownType
) -> Result<(), ErrorCode>

Initiate a graceful shutdown.

receive: the socket is not expecting to receive any more data from the peer. All subsequent read operations on the input-stream associated with this socket will return an End Of Stream indication. Any data still in the receive queue at time of calling shutdown will be discarded.

send: the socket is not expecting to send any more data to the peer. All subsequent write operations on the output-stream associated with this socket will return an error.

both: same effect as receive & send combined.

The shutdown function does not close (drop) the socket.

Typical errors

not-connected: The socket is not in the Connection state. (ENOTCONN)

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/shutdown.html

https://man7.org/linux/man-pages/man2/shutdown.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock/nf-winsock-shutdown

https://man.freebsd.org/cgi/man.cgi?query=shutdown&sektion=2