Function wasi_snapshot_preview1::bindings::wasi::sockets::udp::start_bind

pub fn start_bind(
    this: UdpSocket,
    network: Network,
    local_address: IpSocketAddress
) -> Result<(), ErrorCode>

Bind the socket to a specific network on the provided IP address and port.

If the IP address is zero (0.0.0.0 in IPv4, :: in IPv6), it is left to the implementation to decide which network interface(s) to bind to. If the TCP/UDP port is zero, the socket will be bound to a random free port.

When a socket is not explicitly bound, the first invocation to connect will implicitly bind the socket.

Unlike in POSIX, this function is async. This enables interactive WASI hosts to inject permission prompts.

Typical start errors

address-family-mismatch: The local-address has the wrong address family. (EINVAL)

already-bound: The socket is already bound. (EINVAL)

concurrency-conflict: Another bind or connect operation is already in progress. (EALREADY)

Typical finish errors

ephemeral-ports-exhausted: No ephemeral ports available. (EADDRINUSE, ENOBUFS on Windows)

address-in-use: Address is already in use. (EADDRINUSE)

address-not-bindable: local-address is not an address that the network can bind to. (EADDRNOTAVAIL)

not-in-progress: A bind operation is not in progress.

would-block: Can’t finish the operation, it is still in progress. (EWOULDBLOCK, EAGAIN)

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/bind.html

https://man7.org/linux/man-pages/man2/bind.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock/nf-winsock-bind

https://man.freebsd.org/cgi/man.cgi?query=bind&sektion=2&format=html